# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Policy decision and control plane for agentic systems.",
        "title": "AgentGate API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/internal/integrations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all registered integration definitions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "List integration definitions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.IntegrationDefinition"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a new integration (adapter, resource, etc.).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "Create an integration definition",
                "parameters": [
                    {
                        "description": "Integration definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IntegrationDefinition"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.IntegrationDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/internal/integrations/{integration_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single integration definition by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "Get an integration definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Integration ID",
                        "name": "integration_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.IntegrationDefinition"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes an integration definition.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "Delete an integration definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Integration ID",
                        "name": "integration_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing integration definition.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "summary": "Update an integration definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Integration ID",
                        "name": "integration_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IntegrationDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.IntegrationDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/internal/policy/bundles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all policy bundles, optionally including archived ones.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policy"
                ],
                "summary": "List policy bundles",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "include archived bundles",
                        "name": "include_archived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/policy.Bundle"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new policy bundle.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policy"
                ],
                "summary": "Create a policy bundle",
                "parameters": [
                    {
                        "description": "Bundle definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/policy.Bundle"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/policy.Bundle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/internal/policy/bundles/{bundle_id}/publish": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Publishes a stored policy bundle, making it active.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policy"
                ],
                "summary": "Publish a policy bundle by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bundle ID",
                        "name": "bundle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/internal/policy/bundles/{bundle_id}/validate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates a stored policy bundle.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "policy"
                ],
                "summary": "Validate a policy bundle by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bundle ID",
                        "name": "bundle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.IntegrationDefinition": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "expected_surfaces": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Surface"
                    }
                },
                "health": {
                    "$ref": "#/definitions/types.IntegrationHealth"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "matched_adapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.IntegrationMatchedAdapter"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "types.Surface": {
            "type": "string",
            "enum": [
                "input",
                "runtime",
                "resource"
            ],
            "x-enum-varnames": [
                "SurfaceInput",
                "SurfaceRuntime",
                "SurfaceResource"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Roles: adapter, operator, admin.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}